+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/Abstractions/AuthorizedRequestObjectHandler.cs
@@ -1,11 +1,11 @@
 ﻿namespace InWords.WebApi.Services.Abstractions
 {
-    public class AuthorizedRequestObjectHandler<TQuery, TResult, TContext> :
+    public class AuthReqHandler<TQuery, TResult, TContext> :
         ContextRequestHandler<AuthorizedRequestObject<TQuery, TResult>, TResult, TContext>
         where TQuery : new()
         where TResult : new()
     {
-        public AuthorizedRequestObjectHandler(TContext context) : base(context)
+        public AuthReqHandler(TContext context) : base(context)
         {
 
         }