+++ WebApi/InWords.WebApi/InWords.WebApi/releasenotes.md
@@ -1,5 +1,6 @@
 ﻿### Release notes
 
+- [ ] INW-132 all levels request
 - [x] INW-147 fix history dubles
 - [x] INW-148 remove redundant pipeline artefacts
 - [x] INW-149 update default router